agents/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.74
langchain-groq==0.3.7
langsmith==0.4.16
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.2
packaging==25.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.3
tenacity==9.1.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.24.0
